rename: InstructionDG.h -> .hpp
